apps/api/pyproject.toml
[project]
name = "hhackathon-api"
version = "0.1.0"
description = "FastAPI backend for the Vision Navigation assistant"
requires-python = ">=3.11"
authors = [{ name = "Vision Navigation", email = "team@example.com" }]
dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.1",
  "livekit-api>=0.7.2",
  "httpx>=0.27.2",
  "redis>=5.2.0",
  "numpy>=2.1.2",
  "opencv-python>=4.10.0.84",
  "supervision>=0.22.0",
  "ultralytics>=8.3.25",
  "livekit-agents>=0.9.0",
  "google-cloud-texttospeech>=2.16.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.7.2",
  "mypy>=1.11.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = false

